timout must be a uin64_t

+++ inc/connman/client_set.h
@@ -52,7 +52,7 @@ unsigned int get_max_clients(client_set* cls);
 int reset_max_clients(client_set* cls, unsigned int max_clients);
 
 // this function returns NULL, only when the shutdown is called for the client_set OR if you timed out
-stream* reserve_client(client_set* cls, unsigned int timeout_in_secs);
+stream* reserve_client(client_set* cls, uint64_t timeout_in_secs);
 
 // return a reserved client, this function will close and destroy the client stream if it has errored OR force_destroy_client_stream flag was set
 void return_client(client_set* cls, stream* strm, int force_destroy_client_stream);
